+++ CHANGELOG.md
@@ -31,6 +31,10 @@ Our versioning strategy is as follows:
 
 * `[content]` Add Unified UI origin to allowed origins ([#424](https://github.com/Sitecore/content-sdk/pull/424))
 
+### 🎉 New Features & Improvements
+
+* `[nextjs]` Support for redirect items redirects in redirect proxy (more redirects!) ([#433](https://github.com/Sitecore/content-sdk/pull/433))
+
 ## 2.0.1
 
 ### ✨ Bug fixes